+++ webstepper/CHANGELOG.md
@@ -9,6 +9,8 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### ✨ Enhancements
 
+- Added the MemoryViz logo as a favicon
+
 ### 🐛 Bug fixes
 
 ### 🚨 Breaking changes